requirement.txt
easyocr==1.7.2
filelock==3.16.1
fsspec==2024.10.0
imageio==2.36.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==2.0.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
pyclipper==1.3.0.post6
pydicom==2.4.4
pytesseract==0.3.13
python-bidi==0.6.3
PyYAML==6.0.2
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.6
sympy==1.13.3
tifffile==2024.8.30
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
